+++ controllers/async/mock_AsyncStep_test.go
@@ -5,7 +5,7 @@ package async
 import (
 	context "context"
 
-	v2 "github.com/cloudogu/k8s-dogu-operator/v2/api/v2"
+	v2 "github.com/cloudogu/k8s-dogu-operator/v3/api/v2"
 	mock "github.com/stretchr/testify/mock"
 )