+++ controllers/mock_execPod_test.go
@@ -6,7 +6,7 @@ import (
 	bytes "bytes"
 	context "context"
 
-	exec "github.com/cloudogu/k8s-dogu-operator/v2/controllers/exec"
+	exec "github.com/cloudogu/k8s-dogu-operator/v3/controllers/exec"
 	mock "github.com/stretchr/testify/mock"
 
 	types "k8s.io/apimachinery/pkg/types"